Fix Exception usage
In ruby Exception is the root class of the hierarchy, it should not be
caught most of the time, as that prevents terminating the program with
kill or Ctrl-C. "raise" without a type specified catches StandardError,
that's what is needed most of the time.

On the similar reason Exception should not be thrown directly, "raise"
with a string parameter creates RuntimeError with it.